[Evo] ref #3434 - Add rights through all the app

See merge request internet/stela-poc!120

+++ acte-service/src/main/java/fr/sictiam/stela/acteservice/controller/ActeDraftRestController.java
@@ -3,11 +3,13 @@ package fr.sictiam.stela.acteservice.controller;
 import fr.sictiam.stela.acteservice.model.Acte;
 import fr.sictiam.stela.acteservice.model.ActeMode;
 import fr.sictiam.stela.acteservice.model.LocalAuthority;
+import fr.sictiam.stela.acteservice.model.Right;
 import fr.sictiam.stela.acteservice.model.ui.ActeDraftUI;
 import fr.sictiam.stela.acteservice.model.ui.CustomValidationUI;
 import fr.sictiam.stela.acteservice.model.ui.DraftUI;
 import fr.sictiam.stela.acteservice.service.DraftService;
 import fr.sictiam.stela.acteservice.service.LocalAuthorityService;
+import fr.sictiam.stela.acteservice.service.util.RightUtils;
 import fr.sictiam.stela.acteservice.validation.ValidationUtil;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -22,6 +24,7 @@ import java.io.IOException;
 import java.util.Collections;
 import java.util.List;
 import java.util.Optional;
+import java.util.Set;
 
 @RestController
 @RequestMapping("/api/acte")
@@ -40,8 +43,12 @@ public class ActeDraftRestController {
 
     @GetMapping("/draft/{mode}")
     public ResponseEntity<Acte> newDraft(
+            @RequestAttribute("STELA-Current-Profile-Rights") Set<Right> rights,
             @RequestAttribute("STELA-Current-Local-Authority-UUID") String currentLocalAuthUuid,
             @PathVariable ActeMode mode) {
+        if (!RightUtils.hasRight(rights, Collections.singletonList(Right.ACTES_DEPOSIT))) {
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
         LocalAuthority currentLocalAuthority = localAuthorityService.getByUuid(currentLocalAuthUuid);
         Acte acte = draftService.newDraft(currentLocalAuthority, mode);
         return new ResponseEntity<>(acte, HttpStatus.OK);
@@ -49,41 +56,63 @@ public class ActeDraftRestController {
 
     @GetMapping("/draft/batch")
     public ResponseEntity<DraftUI> newBatchedDraft(
+            @RequestAttribute("STELA-Current-Profile-Rights") Set<Right> rights,
             @RequestAttribute("STELA-Current-Local-Authority-UUID") String currentLocalAuthUuid) {
+        if (!RightUtils.hasRight(rights, Collections.singletonList(Right.ACTES_DEPOSIT))) {
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
         LocalAuthority currentLocalAuthority = localAuthorityService.getByUuid(currentLocalAuthUuid);
         DraftUI draft = draftService.newBatchedDraft(currentLocalAuthority);
         return new ResponseEntity<>(draft, HttpStatus.OK);
     }
 
     @GetMapping("/drafts")
-    public ResponseEntity<List<DraftUI>> getDrafts() {
+    public ResponseEntity<List<DraftUI>> getDrafts(@RequestAttribute("STELA-Current-Profile-Rights") Set<Right> rights) {
+        if (!RightUtils.hasRight(rights, Collections.singletonList(Right.ACTES_DEPOSIT))) {
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
         List<DraftUI> drafts = draftService.getDraftUIs();
         return new ResponseEntity<>(drafts, HttpStatus.OK);
     }
 
     @DeleteMapping("/drafts")
-    public ResponseEntity<?> deleteDrafts(@RequestBody List<String> uuids) {
+    public ResponseEntity<?> deleteDrafts(@RequestAttribute("STELA-Current-Profile-Rights") Set<Right> rights,
+            @RequestBody List<String> uuids) {
+        if (!RightUtils.hasRight(rights, Collections.singletonList(Right.ACTES_DEPOSIT))) {
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
         draftService.deleteDrafts(uuids);
         return new ResponseEntity<>(HttpStatus.OK);
     }
 
     @GetMapping("/drafts/{uuid}")
-    public ResponseEntity<DraftUI> getDraftByUuid(@PathVariable String uuid) {
+    public ResponseEntity<DraftUI> getDraftByUuid(@RequestAttribute("STELA-Current-Profile-Rights") Set<Right> rights,
+            @PathVariable String uuid) {
+        if (!RightUtils.hasRight(rights, Collections.singletonList(Right.ACTES_DEPOSIT))) {
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
         DraftUI draft = draftService.getDraftActesUI(uuid);
         return new ResponseEntity<>(draft, HttpStatus.OK);
     }
 
     @GetMapping("/drafts/{draftUuid}/{uuid}")
-    public ResponseEntity<Acte> getActeDraftByUuid(@PathVariable String uuid) {
+    public ResponseEntity<Acte> getActeDraftByUuid(@RequestAttribute("STELA-Current-Profile-Rights") Set<Right> rights,
+            @PathVariable String uuid) {
+        if (!RightUtils.hasRight(rights, Collections.singletonList(Right.ACTES_DEPOSIT))) {
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
         Acte acte = draftService.getActeDraftByUuid(uuid);
         return new ResponseEntity<>(acte, HttpStatus.OK);
     }
 
     @PutMapping("/drafts/{draftUuid}/{uuid}")
     ResponseEntity<String> saveActeDraft(
+            @RequestAttribute("STELA-Current-Profile-Rights") Set<Right> rights,
             @RequestAttribute("STELA-Current-Local-Authority-UUID") String currentLocalAuthUuid,
             @RequestBody Acte acte) {
-        // TODO Retrieve current local authority
+        if (!RightUtils.hasRight(rights, Collections.singletonList(Right.ACTES_DEPOSIT))) {
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
         LocalAuthority currentLocalAuthority = localAuthorityService.getByUuid(currentLocalAuthUuid);
         Acte result = draftService.saveActeDraft(acte, currentLocalAuthority);
         return new ResponseEntity<>(result.getUuid(), HttpStatus.OK);
@@ -91,7 +120,11 @@ public class ActeDraftRestController {
 
     @PostMapping("/drafts/{draftUuid}")
     ResponseEntity<?> submitDraft(@PathVariable String draftUuid,
+            @RequestAttribute("STELA-Current-Profile-Rights") Set<Right> rights,
             @RequestAttribute("STELA-Current-Profile-UUID") String profileUuid) {
+        if (!RightUtils.hasRight(rights, Collections.singletonList(Right.ACTES_DEPOSIT))) {
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
         Optional opt = draftService.sumitDraft(draftUuid, profileUuid);
         if (opt.isPresent())
             return new ResponseEntity<>(opt.get(), HttpStatus.BAD_REQUEST);
@@ -99,20 +132,32 @@ public class ActeDraftRestController {
     }
 
     @PatchMapping("/drafts/{draftUuid}")
-    ResponseEntity<?> upddateDraftFields(@RequestBody DraftUI draftUI) {
+    ResponseEntity<?> upddateDraftFields(@RequestAttribute("STELA-Current-Profile-Rights") Set<Right> rights,
+            @RequestBody DraftUI draftUI) {
+        if (!RightUtils.hasRight(rights, Collections.singletonList(Right.ACTES_DEPOSIT))) {
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
         draftService.upddateDraftFields(draftUI);
         return new ResponseEntity<>(HttpStatus.CREATED);
     }
 
     @DeleteMapping("/drafts/{draftUuid}")
-    ResponseEntity<?> deleteDraft(@PathVariable String draftUuid) {
+    ResponseEntity<?> deleteDraft(@RequestAttribute("STELA-Current-Profile-Rights") Set<Right> rights,
+            @PathVariable String draftUuid) {
+        if (!RightUtils.hasRight(rights, Collections.singletonList(Right.ACTES_DEPOSIT))) {
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
         draftService.deleteDrafts(Collections.singletonList(draftUuid));
         return new ResponseEntity<>(HttpStatus.CREATED);
     }
 
     @PostMapping("/drafts/{draftUuid}/{uuid}")
     ResponseEntity<?> submitActeDraft(@PathVariable String uuid,
+            @RequestAttribute("STELA-Current-Profile-Rights") Set<Right> rights,
             @RequestAttribute("STELA-Current-Profile-UUID") String profileUuid) {
+        if (!RightUtils.hasRight(rights, Collections.singletonList(Right.ACTES_DEPOSIT))) {
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
         Acte acteDraft = draftService.getActeDraftByUuid(uuid);
         List<ObjectError> errors = ValidationUtil.validateActe(acteDraft);
         if (!errors.isEmpty()) {
@@ -127,24 +172,35 @@ public class ActeDraftRestController {
 
     @PostMapping("/drafts/{uuid}/newActe")
     public ResponseEntity<ActeDraftUI> newActeForDraft(
+            @RequestAttribute("STELA-Current-Profile-Rights") Set<Right> rights,
             @RequestAttribute("STELA-Current-Local-Authority-UUID") String currentLocalAuthUuid,
             @PathVariable String uuid) {
-        // TODO Retrieve current local authority
+        if (!RightUtils.hasRight(rights, Collections.singletonList(Right.ACTES_DEPOSIT))) {
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
         LocalAuthority currentLocalAuthority = localAuthorityService.getByUuid(currentLocalAuthUuid);
         ActeDraftUI acte = draftService.newActeForDraft(uuid, currentLocalAuthority);
         return new ResponseEntity<>(acte, HttpStatus.OK);
     }
 
     @DeleteMapping("/drafts/{uuid}/{acteUuid}")
-    public ResponseEntity<?> deleteActeDraft(@PathVariable("acteUuid") String acteUuid) {
+    public ResponseEntity<?> deleteActeDraft(@RequestAttribute("STELA-Current-Profile-Rights") Set<Right> rights,
+            @PathVariable("acteUuid") String acteUuid) {
+        if (!RightUtils.hasRight(rights, Collections.singletonList(Right.ACTES_DEPOSIT))) {
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
         draftService.deleteActeDraftByUuid(acteUuid);
         return new ResponseEntity<>(HttpStatus.OK);
     }
 
     @PutMapping("/drafts/{draftUuid}/{uuid}/leave")
     public ResponseEntity<?> leaveActeDraft(
+            @RequestAttribute("STELA-Current-Profile-Rights") Set<Right> rights,
             @RequestAttribute("STELA-Current-Local-Authority-UUID") String currentLocalAuthUuid,
             @RequestBody Acte acte) {
+        if (!RightUtils.hasRight(rights, Collections.singletonList(Right.ACTES_DEPOSIT))) {
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
         LocalAuthority currentLocalAuthority = localAuthorityService.getByUuid(currentLocalAuthUuid);
         draftService.leaveActeDraft(acte, currentLocalAuthority);
         return new ResponseEntity<>(HttpStatus.OK);
@@ -152,8 +208,12 @@ public class ActeDraftRestController {
 
     @PostMapping("/drafts/{draftUuid}/{uuid}/file")
     public ResponseEntity<Acte> saveActeDraftFile(
+            @RequestAttribute("STELA-Current-Profile-Rights") Set<Right> rights,
             @RequestAttribute("STELA-Current-Local-Authority-UUID") String currentLocalAuthUuid,
             @PathVariable String uuid, @RequestParam("file") MultipartFile file) {
+        if (!RightUtils.hasRight(rights, Collections.singletonList(Right.ACTES_DEPOSIT))) {
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
         LocalAuthority currentLocalAuthority = localAuthorityService.getByUuid(currentLocalAuthUuid);
         try {
             Acte acte = draftService.saveActeDraftFile(uuid, file, currentLocalAuthority);
@@ -166,8 +226,12 @@ public class ActeDraftRestController {
 
     @PostMapping("/drafts/{draftUuid}/{uuid}/annexe")
     public ResponseEntity<Acte> saveActeDraftAnnexe(
+            @RequestAttribute("STELA-Current-Profile-Rights") Set<Right> rights,
             @RequestAttribute("STELA-Current-Local-Authority-UUID") String currentLocalAuthUuid,
             @PathVariable String uuid, @RequestParam("file") MultipartFile file) {
+        if (!RightUtils.hasRight(rights, Collections.singletonList(Right.ACTES_DEPOSIT))) {
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
         LocalAuthority currentLocalAuthority = localAuthorityService.getByUuid(currentLocalAuthUuid);
         try {
             Acte acte = draftService.saveActeDraftAnnexe(uuid, file, currentLocalAuthority);
@@ -179,31 +243,52 @@ public class ActeDraftRestController {
     }
 
     @PutMapping("/drafts/{draftUuid}/{acteUuid}/file/type/{uuid}")
-    public ResponseEntity updateFileAttachmentType(@PathVariable String acteUuid, @PathVariable String uuid) {
+    public ResponseEntity updateFileAttachmentType(@RequestAttribute("STELA-Current-Profile-Rights") Set<Right> rights,
+            @PathVariable String acteUuid, @PathVariable String uuid) {
+        if (!RightUtils.hasRight(rights, Collections.singletonList(Right.ACTES_DEPOSIT))) {
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
         draftService.updateActeFileAttachmentType(acteUuid, uuid);
         return new ResponseEntity<>(HttpStatus.OK);
     }
 
     @PutMapping("/drafts/{draftUuid}/{acteUuid}/annexe/{annexeUuid}/type/{uuid}")
-    public ResponseEntity updateAnnexeAttachmentType(@PathVariable String annexeUuid, @PathVariable String uuid) {
+    public ResponseEntity updateAnnexeAttachmentType(
+            @RequestAttribute("STELA-Current-Profile-Rights") Set<Right> rights,
+            @PathVariable String annexeUuid, @PathVariable String uuid) {
+        if (!RightUtils.hasRight(rights, Collections.singletonList(Right.ACTES_DEPOSIT))) {
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
         draftService.updateAttachmentType(annexeUuid, uuid);
         return new ResponseEntity<>(HttpStatus.OK);
     }
 
     @DeleteMapping("/drafts/{draftUuid}/types")
-    public ResponseEntity removeAttachmentTypes(@PathVariable String draftUuid) {
+    public ResponseEntity removeAttachmentTypes(@RequestAttribute("STELA-Current-Profile-Rights") Set<Right> rights,
+            @PathVariable String draftUuid) {
+        if (!RightUtils.hasRight(rights, Collections.singletonList(Right.ACTES_DEPOSIT))) {
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
         draftService.removeAttachmentTypes(draftUuid);
         return new ResponseEntity<>(HttpStatus.OK);
     }
 
     @DeleteMapping("/drafts/{draftUuid}/{uuid}/file")
-    public ResponseEntity<?> deleteActeDraftFile(@PathVariable String uuid) {
+    public ResponseEntity<?> deleteActeDraftFile(@RequestAttribute("STELA-Current-Profile-Rights") Set<Right> rights,
+            @PathVariable String uuid) {
+        if (!RightUtils.hasRight(rights, Collections.singletonList(Right.ACTES_DEPOSIT))) {
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
         draftService.deleteActeDraftFile(uuid);
         return new ResponseEntity<>(HttpStatus.OK);
     }
 
     @DeleteMapping("/drafts/{draftUuid}/{acteUuid}/annexe/{uuid}")
-    public ResponseEntity<?> deleteActeDraftAnnexe(@PathVariable String acteUuid, @PathVariable String uuid) {
+    public ResponseEntity<?> deleteActeDraftAnnexe(@RequestAttribute("STELA-Current-Profile-Rights") Set<Right> rights,
+            @PathVariable String acteUuid, @PathVariable String uuid) {
+        if (!RightUtils.hasRight(rights, Collections.singletonList(Right.ACTES_DEPOSIT))) {
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
         draftService.deleteActeDraftAnnexe(acteUuid, uuid);
         return new ResponseEntity<>(HttpStatus.OK);
     }
